refactor: add interface implementation
