+++ src/main/java/racingcar/model/Cars.java
@@ -1,5 +1,7 @@
 package racingcar.model;
 
+import racingcar.util.RandomRange;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -14,5 +16,11 @@ public class Cars {
 
     public List<Car> getCars() {
         return cars;
+    }
+
+    public void runRace() {
+        for (Car car : cars) {
+            car.move(RandomRange.getRandomValue());
+        }
     }
 }
