+++ src/main/java/racingcar/model/MovingStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.model;
+
+public interface MovingStrategy {
+    boolean matchMovementCondition(int value);
+}
