+++ src/main/java/racingcar/util/RandomRange.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.util;
 
 import java.util.Random;
