+++ src/test/java/racingcar/util/RandomRangeTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
